Fixes #12 , #18 , #19 and #25
I improved File support so now torrents can be opened in place.
I fixed the issue where the keyboard covers the text field.
Fixed a parsing issue with Deluge 2.0.3

+++ Deluge Remote/Sources/Models/Deluge/SessionStatus.swift
@@ -27,7 +27,7 @@ struct SessionStatus: Decodable {
     /// but only parts of the actual files to be downloaded.
     let total_payload_upload: Int
 
-    let upload_rate: Int
+    let upload_rate: Double
 
     /// payload_upload_rate is the rate of the payload down- and upload only.
     let payload_upload_rate: Int
